requirements.txt
django-autoslug==1.9.8
django-countries==7.2.1
django-cities==0.6
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
django-filter==21.1
gunicorn==20.1.0
ipdb==0.13.9
markdown==3.3.6
Pillow==8.4.0
psycopg2-binary==2.9.2
